+++ packages/task-manager/src/main.tsx
@@ -1,7 +1,6 @@
-import React from "react";
 import ReactDOM from "react-dom/client";
 import { container } from "storion";
-import { StoreProvider } from "storion/react";
+import { StoreProvider, StrictMode } from "storion/react";
 import { devtoolsMiddleware } from "storion/devtools";
 import { mountDevtoolsPanel } from "storion/devtools-panel";
 import App from "./App";
@@ -28,9 +27,9 @@ mountDevtoolsPanel({
 });
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
-  <React.StrictMode>
+  <StrictMode>
     <StoreProvider container={app}>
       <App />
     </StoreProvider>
-  </React.StrictMode>
+  </StrictMode>
 );